Move CrashDump to pocketmine\crash namespace
